- webAppProject faq 모듈 refactoring 완료

+++ src/main/java/com/dev/webApp/domain/vo/FaqVO.java
@@ -15,7 +15,7 @@ import java.util.Date;
 @AllArgsConstructor
 public class FaqVO {
 
-    Long faqNo;
+    Integer faqNo;
 
     String title;
